+++ packages/ui/certd-client/src/views/certd/pipeline/certd-form/use.tsx
@@ -152,6 +152,13 @@ export function useCertPipelineCreator({ formWrapperRef }: { formWrapperRef: Ref
     });
 
     const initialForm = req.initialForm || {};
+    const pluginSysConfig = await pluginStore.getPluginConfig({ name: certPlugin.name, type: "builtIn" });
+    if (pluginSysConfig?.sysSetting?.input) {
+      //设置系统默认值，必须放这里设置，保证随时应用最新的系统值
+      for (const key in pluginSysConfig.sysSetting?.input) {
+        initialForm.input[key] = pluginSysConfig.sysSetting?.input[key];
+      }
+    }
     initialForm.type = certPlugin.name;
     const applyTemplates = reactive<any[]>([]);
     const selectedTemplateId = ref<number | null>(null);
@@ -556,12 +563,12 @@ export function useCertPipelineCreator({ formWrapperRef }: { formWrapperRef: Ref
 
     //设置系统初始值
     const initialForm: any = { input: {} };
-    const pluginSysConfig = await pluginStore.getPluginConfig({ name: req.pluginName, type: "builtIn" });
-    if (pluginSysConfig.sysSetting?.input) {
-      for (const key in pluginSysConfig.sysSetting?.input) {
-        initialForm.input[key] = pluginSysConfig.sysSetting?.input[key];
-      }
-    }
+    // const pluginSysConfig = await pluginStore.getPluginConfig({ name: req.pluginName, type: "builtIn" });
+    // if (pluginSysConfig.sysSetting?.input) {
+    //   for (const key in pluginSysConfig.sysSetting?.input) {
+    //     initialForm.input[key] = pluginSysConfig.sysSetting?.input[key];
+    //   }
+    // }
     async function doSubmit({ form }: any) {
       // const certDetail = readCertDetail(form.cert.crt);
       // 添加certd pipeline