+++ packages/ui/certd-client/src/views/sys/plugin/config-common.vue
@@ -69,7 +69,7 @@
 <script lang="ts" setup>
 import AccessSelector from "/@/views/certd/access/access-selector/index.vue";
 import { reactive, ref } from "vue";
-import { CommPluginConfig, GetCommPluginConfigs, SaveCommPluginConfigs, GetPluginByName } from "/@/views/sys/plugin/api";
+import { CommPluginConfig, GetCommPluginConfigs, SaveCommPluginConfigs, getPluginSetting } from "/@/views/sys/plugin/api";
 import { merge } from "lodash-es";
 import { notification } from "ant-design-vue";
 import { usePluginConfig } from "./use-config";
@@ -113,7 +113,7 @@ const onFinishFailed = (errorInfo: any) => {
 const { openConfigDialog } = usePluginConfig();
 
 async function doPluginConfig() {
-  const certApplyInfo = await GetPluginByName("CertApply");
+  const certApplyInfo = await getPluginSetting("CertApply");
   await openConfigDialog({
     row: certApplyInfo,
     onSuccess: async () => {