+++ packages/ui/certd-server/src/modules/auto/fix/common-eab-to-acme-account-fix.ts
@@ -93,6 +93,7 @@ export class CommonEabToAcmeAccountFix {
             ...certApplyConfig.sysSetting,
             input,
           },
+          type: "builtIn",
         });
       }
       return true;